requirements.txt
configtune==0.0.8
cycler==0.10.0
deap==1.3.0
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
pandas==0.25.2
pudb==2019.1
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
scikit-learn==0.21.3
scikit-optimize==0.5.2
scipy==1.3.1
six==1.12.0
sklearn==0.0
torch==1.3.0
tqdm==4.37.0
urwid==2.0.1
